move refresh metadata to metadata package
